audio_core/codec: Resolve truncation warnings within DecodeADPCM
The assignments here were performing an implicit truncation from int to
s16. Make it explicit that this is desired behavior.

+++ src/audio_core/codec.cpp
@@ -72,8 +72,8 @@ StereoBuffer16 DecodeADPCM(const u8* const data, const std::size_t sample_count,
         }
     }
 
-    state.yn1 = yn1;
-    state.yn2 = yn2;
+    state.yn1 = static_cast<s16>(yn1);
+    state.yn2 = static_cast<s16>(yn2);
 
     return ret;
 }
